# api-specs/openapi-spec-20251021-190000.yaml
openapi: 3.1.0
info:
  title: Decentralized Autonomous Forum API
  version: 1.0.0
  description: |
    RESTful API for the Decentralized Autonomous Forum platform. Features:
    - Point economy with crypto rewards
    - User authentication (email + OAuth2)
    - Post/comment creation with IPFS storage
    - Reputation-based moderation
    - BNB Chain integration

    **Base URL**: `https://api.forum.example.com/v1`

    **Authentication**: Session-based (HttpOnly cookies) + JWT for API clients

    **Rate Limiting**:
    - Anonymous: 100 req/hour
    - Authenticated: 1000 req/hour
    - Authenticated (Level 3+): 5000 req/hour

  contact:
    name: API Support
    email: support@forum.example.com

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.forum.example.com/v1
    description: Production server
  - url: https://staging-api.forum.example.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Local development server
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: Session cookie (HttpOnly, Secure, SameSite=Lax)

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for API clients

# ============================================================================
# REUSABLE SCHEMAS
# ============================================================================
  schemas:
...
    Comment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        post_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        author:
          $ref: '#/components/schemas/UserSummary'
        parent_id:
          type: integer
          format: int64
          nullable: true
        body:
          type: string
          minLength: 1
          maxLength: 2000
        body_html:
          type: string
        like_count:
          type: integer
          minimum: 0
        depth:
          type: integer
          minimum: 0
          maximum: 2
        replies:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        created_at:
          type: string
          format: date-time
        edited_at:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - post_id
        - user_id
        - author
        - body
        - like_count
        - depth
        - created_at
...
    CreateCommentRequest:
      type: object
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 2000
        parent_id:
          type: integer
          format: int64
          nullable: true
      required:
        - body

    # ========== Other Schemas ==========
    UserSummary:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
        level:
          type: string
        is_verified:
          type: boolean
...
paths:
...
  /posts/{post_id}/comments:
    get:
      tags:
        - Comments
      summary: Get post comments
      description: Get all comments for a post (nested structure)
      operationId: getPostComments
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Comment'

    post:
      tags:
        - Comments
      summary: Create comment
      description: Add a comment to a post (costs 2 points)
      operationId: createComment
      security:
        - SessionAuth: []
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '403':
          description: Insufficient points or rate limit exceeded
  /comments/{comment_id}:
    delete:
      tags:
        - Comments
      summary: Delete comment
      description: Delete own comment (points not refunded)
      operationId: deleteComment
      security:
        - SessionAuth: []
      parameters:
        - name: comment_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Comment deleted
        '403':
          description: Cannot delete comment
        '404':
          description: Comment not found

  # ========== Like Endpoints ==========
  /posts/{post_id}/like:
    post:
      tags:
        - Likes
      summary: Like a post
      description: Like a post (costs 1 point)
      operationId: likePost
      security:
        - SessionAuth: []
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Post liked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Post liked"
                  like_count:
                    type: integer
                    example: 235
        '403':
          description: Insufficient points, already liked, or cannot self-like
        '404':
          description: Post not found

    delete:
      tags:
        - Likes
      summary: Unlike a post
      description: Remove like from post (point refunded)
      operationId: unlikePost
      security:
        - SessionAuth: []
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Post unliked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Post unliked"
                  like_count:
                    type: integer
                    example: 234

  # ========== Point/Transaction Endpoints ==========
...